Moved several files and configured the CMakeLists.txt to correctly install everything.
* Added missing python files through the setup.py
* Automatically downloads the model and install it.
